refactor: Separate define class file from index.ts.
